telegrambot/requirements.txt
Flask==2.2.2
google-cloud-pubsub==2.13.11
google-cloud-storage==2.7.0
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.0
